update and test server search

+++ tests/test_search_server.py
@@ -1,6 +1,7 @@
 
 from kthread import KThread
 from os.path import join, dirname
+from os import remove
 from unittest import TestCase
 
 from ariesk.search_server import (
@@ -9,6 +10,7 @@ from ariesk.search_server import (
 )
 
 GRID_COVER = join(dirname(__file__), 'small_grid_cover.sqlite')
+KMER_TABLE = join(dirname(__file__), 'small_31mer_table.csv')
 
 PORT = 5432
 KMER_31 = 'AATACGTCCGGAGTATCGACGCACACATGGT'
@@ -29,7 +31,29 @@ class TestSearchServer(TestCase):
             client.send_shutdown()
             self.assertIn(KMER_31, results)
         finally:
-            server_thread.terminate()
+            if server_thread.is_alive():
+                server_thread.terminate()
+
+    def test_search_server_file(self):
+        client = SearchClient(PORT)
+        server_thread = KThread(target=run_server)
+        outfile = 'temp.test_outfile.csv'
+        try:
+            server_thread.start()
+            client.search(
+                KMER_TABLE, 0.001, 0.1,
+                result_file=outfile, query_type='file'
+            )
+            client.send_shutdown()
+            results = []
+            with open(outfile) as f:
+                for line in f:
+                    results.append(line.strip().split()[1])
+            remove(outfile)
+            self.assertIn(KMER_31, results)
+        finally:
+            if server_thread.is_alive():
+                server_thread.terminate()
 
     def test_coarse_search_server(self):
         client = SearchClient(PORT)
@@ -40,7 +64,8 @@ class TestSearchServer(TestCase):
             client.send_shutdown()
             self.assertGreaterEqual(len(results), 1)
         finally:
-            server_thread.terminate()
+            if server_thread.is_alive():
+                server_thread.terminate()
 
     def test_search_server_no_inner(self):
         client = SearchClient(PORT)
@@ -51,4 +76,5 @@ class TestSearchServer(TestCase):
             client.send_shutdown()
             self.assertIn(KMER_31, results)
         finally:
-            server_thread.terminate()
+            if server_thread.is_alive():
+                server_thread.terminate()
